requirements.txt
fastapi[standard]
google-genai
pymupdf
python-multipart
pillow